requirements.txt
transformers
datasets
sentence-transformers
faiss-cpu
mlflow
fastapi
uvicorn
torch
